# Optimize colorize using matmul and inplace operations