@hoxbro 2026-03-19
Removed the einsum calculation as I could not find significant improvement, and was very inconsistent based on number of categories. Also changed the data to be float32, in my opinion the higher precision is not needed.  

<details><summary>Performance code</summary>
<p>

``` python
import time
import numpy as np
import pandas as pd
import datashader as ds
import datashader.transfer_functions as tf

N = int(10e6)
C = 20


class Profile:
    def __init__(self, output_file):
        import cProfile

        self.profiler = cProfile.Profile()
        self.output_file = output_file

    def __enter__(self):
        self.profiler.enable()
        return self.profiler

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.profiler.disable()
        self.profiler.dump_stats(self.output_file)


class LineProfileContext:
    def __init__(self, output_file):
        from line_profiler import LineProfiler

        self.profiler = LineProfiler()
        self.output_file = output_file
        self.functions_to_profile = []

    def add_function(self, func):
        """Add a function to be profiled line-by-line"""
        self.profiler.add_function(func)
        self.functions_to_profile.append(func)
        return func

    def __enter__(self):
        self.profiler.enable_by_count()
        return self

    def __exit__(self, exc_type, exc_val, exc_tb):
        self.profiler.disable_by_count()

        self.profiler.dump_stats(self.output_file)
        self.profiler.print_stats()


def gen_data(N=N, C=C):
    np.random.seed(1)
    xy = np.random.randn(int(N), 2)
    c = np.random.choice([chr(65 + i) for i in range(C)], size=N)
    df = pd.DataFrame(xy, columns=["x", "y"])
    df["c"] = pd.Series(c).astype("category")
    return df


def profile(df, size=1000):
    W = H = size
    cvs = ds.Canvas(plot_width=W, plot_height=H)
    agg = cvs.points(df, x="x", y="y", agg=ds.count_cat("c"))

    tf.shade(agg)  # warm up
    pre = time.monotonic()

    # with LineProfileContext("line_profile.lprof") as line_profiler:
    #     # line_profiler.add_function(tf._colorize)
    #     line_profiler.add_function(tf._interpolate_alpha)
    #     tf.shade(agg)

    with Profile(output_file="optional.perf"):
        ds.transfer_functions.shade(agg)
    return time.monotonic() - pre


def main():
    df = gen_data(C=20)
    # profile(df, size=5000)

    results = []
    for c in (1, 5, 10, 20):
        df = gen_data(C=c)
        for s in range(1000, 6000, 1000):
            timing = profile(df, size=s)
            results.append((c, s, timing))
            print(f"dict({c=}, {s=}, {timing=}),")


main()
```

</p>
</details> 

<details><summary>Plotting code</summary>
<p>

``` python
einsum = [  # 2c78c6d
    dict(c=1, s=1000, timing=0.06803519899949606),
    dict(c=1, s=2000, timing=0.2663327520003804),
    dict(c=1, s=3000, timing=0.5744440779999422),
    dict(c=1, s=4000, timing=1.0103413500000897),
    dict(c=1, s=5000, timing=1.5277441300004284),
    dict(c=5, s=1000, timing=0.13711750999937067),
    dict(c=5, s=2000, timing=0.5479344109999147),
    dict(c=5, s=3000, timing=1.1772052269998312),
    dict(c=5, s=4000, timing=2.1423583199994027),
    dict(c=5, s=5000, timing=3.1938455540002906),
    dict(c=10, s=1000, timing=0.17679083300026832),
    dict(c=10, s=2000, timing=0.7543239710003036),
    dict(c=10, s=3000, timing=1.587601693000579),
    dict(c=10, s=4000, timing=3.1705693840003732),
    dict(c=10, s=5000, timing=4.292662657000619),
    dict(c=20, s=1000, timing=0.3295214039999337),
    dict(c=20, s=2000, timing=1.266628225000204),
    dict(c=20, s=3000, timing=3.5860006670000075),
    dict(c=20, s=4000, timing=13.614561829000195),
    dict(c=20, s=5000, timing=9.398881717999757),
]

matrix = [  # 7bff52a
    dict(c=1, s=1000, timing=0.0629920419996779),
    dict(c=1, s=2000, timing=0.24183893400004308),
    dict(c=1, s=3000, timing=0.5292455299995709),
    dict(c=1, s=4000, timing=0.9098843479996503),
    dict(c=1, s=5000, timing=1.399802973000078),
    dict(c=5, s=1000, timing=0.1419354019999446),
    dict(c=5, s=2000, timing=0.5491873940000005),
    dict(c=5, s=3000, timing=1.2099411299996063),
    dict(c=5, s=4000, timing=2.1237009349997606),
    dict(c=5, s=5000, timing=3.3163266550000117),
    dict(c=10, s=1000, timing=0.17571856299946376),
    dict(c=10, s=2000, timing=0.6789577050003572),
    dict(c=10, s=3000, timing=1.4855017199997747),
    dict(c=10, s=4000, timing=2.613523547999648),
    dict(c=10, s=5000, timing=4.052620502000536),
    dict(c=20, s=1000, timing=0.23399006699946767),
    dict(c=20, s=2000, timing=0.91144125000028),
    dict(c=20, s=3000, timing=2.026365684000666),
    dict(c=20, s=4000, timing=3.5640572980000798),
    dict(c=20, s=5000, timing=5.543187566000597),
]
float32 = [  # 4f5bdce
    dict(c=1, s=1000, timing=0.058510111999567016),
    dict(c=1, s=2000, timing=0.20919769600004656),
    dict(c=1, s=3000, timing=0.45244464699953824),
    dict(c=1, s=4000, timing=0.7802826970000751),
    dict(c=1, s=5000, timing=1.1908017530004145),
    dict(c=5, s=1000, timing=0.12758468500032905),
    dict(c=5, s=2000, timing=0.4868928870000673),
    dict(c=5, s=3000, timing=1.0767511809999633),
    dict(c=5, s=4000, timing=1.8854512970001451),
    dict(c=5, s=5000, timing=2.9044053600000552),
    dict(c=10, s=1000, timing=0.15367073500055994),
    dict(c=10, s=2000, timing=0.5792317429995819),
    dict(c=10, s=3000, timing=1.2640718610000476),
    dict(c=10, s=4000, timing=2.2208463349998055),
    dict(c=10, s=5000, timing=3.438165036999635),
    dict(c=20, s=1000, timing=0.19284271800006536),
    dict(c=20, s=2000, timing=0.7703899379994255),
    dict(c=20, s=3000, timing=1.655678447000355),
    dict(c=20, s=4000, timing=2.9326316679998854),
    dict(c=20, s=5000, timing=4.665550198999881),
]

pre_rgba = [  # 1caf98e
    dict(c=1, s=1000, timing=0.05790932600029919),
    dict(c=1, s=2000, timing=0.20509067500006495),
    dict(c=1, s=3000, timing=0.4477888579995124),
    dict(c=1, s=4000, timing=0.7889217060001101),
    dict(c=1, s=5000, timing=1.2035187020001104),
    dict(c=5, s=1000, timing=0.12838444999943022),
    dict(c=5, s=2000, timing=0.4901264459995218),
    dict(c=5, s=3000, timing=1.081453621000037),
    dict(c=5, s=4000, timing=1.9430143469999166),
    dict(c=5, s=5000, timing=2.9419500280000648),
    dict(c=10, s=1000, timing=0.16017448599995987),
    dict(c=10, s=2000, timing=0.6139489810002488),
    dict(c=10, s=3000, timing=1.292098817999431),
    dict(c=10, s=4000, timing=2.276405729999169),
    dict(c=10, s=5000, timing=3.548698446999879),
    dict(c=20, s=1000, timing=0.19487345500056108),
    dict(c=20, s=2000, timing=0.7641674759997841),
    dict(c=20, s=3000, timing=1.7333330170004047),
    dict(c=20, s=4000, timing=2.9618597700000464),
    dict(c=20, s=5000, timing=4.6101887670001815),
]

main = [  #  2afcc79
    dict(c=1, s=1000, timing=0.0772107610000603),
    dict(c=1, s=2000, timing=0.3082689709999613),
    dict(c=1, s=3000, timing=0.7179236020001554),
    dict(c=1, s=4000, timing=1.228880463000678),
    dict(c=1, s=5000, timing=1.8922187200005283),
    dict(c=5, s=1000, timing=0.19654757299940684),
    dict(c=5, s=2000, timing=0.7806584370000564),
    dict(c=5, s=3000, timing=1.7628902300002665),
    dict(c=5, s=4000, timing=3.0790194150004027),
    dict(c=5, s=5000, timing=4.802155132000735),
    dict(c=10, s=1000, timing=0.24441086200022255),
    dict(c=10, s=2000, timing=0.9614489120003782),
    dict(c=10, s=3000, timing=2.153044267000041),
    dict(c=10, s=4000, timing=3.8028123309995863),
    dict(c=10, s=5000, timing=5.870340117000524),
    dict(c=20, s=1000, timing=0.3279168019998906),
    dict(c=20, s=2000, timing=1.3417445670002053),
    dict(c=20, s=3000, timing=2.95393913999942),
    dict(c=20, s=4000, timing=5.206289214000208),
    dict(c=20, s=5000, timing=8.059511065000152),
]

import hvplot.pandas
import pandas as pd

df_ref = pd.DataFrame(main)


def fn(x):
    df = pd.DataFrame(eval(x))
    df["timing"] = df_ref["timing"] / df["timing"]
    return df.hvplot.bar(x="s", y="timing", by="c", title=x).opts(show_grid=True)


(fn("einsum") + fn("matrix") + fn("float32") + fn("pre_rgba")).cols(2)
```

</p>
</details> 

<img width="1400" height="630" alt="bokeh_plot" src="https://github.com/user-attachments/assets/6aa2cd9c-4c91-4e44-bb44-cec9f748c8d3" />

--- 
Also verified that tests still work on GPU:

<img width="1083" height="320" alt="image" src="https://github.com/user-attachments/assets/8f8293ec-4379-4580-9a27-39922052ed45" />